// src/utils/enums/form-message.ts
export enum FORM_MESSAGE {
  NAME = "O campo nome deve ser preenchido!",
  CPF = "O campo CPF deve ser preenchido!",
  EMAIL = "O campo e-mail deve ser preenchido!",
  PASSWORD = "O campo senha deve ser preenchido!",
  CONFIRM_PASSWORD = "O campo confirmar senha seve ser prenchido!",
  PASSWORD_CONFIRM_MATCH = "As senhas devem ser iguais!",
  CPF_SIZE = "O CPF deve ter 11 dígitos!",
  CPF_VALID = "O CPF deve ser válido!",
  EMAIL_FORMAT = "O e-mail deve ser válido!",
  PASSWORD_SIZE = "A senha deve ter no mínimo 8 dígitos!",
  PASSWORD_RULE = "A senha não atingiu os requisitos mínimos!",
  IMAGE_FORMAT = "A imagem deve estar nos formatos: PNG, JPG, JPEG"
}